Rename environment -> context
